fix: address review — split RewardVerificationRuntime to its own file
Moves the runtime class out of RewardVerificationManager.kt so each file
holds a single top-level type, mirroring the existing test file split
between RewardVerificationManagerTest and RewardVerificationRuntimeTest.
